Follow-ups to the astral-plane repair: unit tests for mosh_mbrtoc32's
malformed-input contract, a native-CI regression test proving an astral
scalar takes the no-prediction branch in PredictionEngine (the original
truncation bug site), and the mosh_char_t typedef removal finishing the
canonical char32_t scalar type.

Sponsored-By: Netflix

* mosh-prediction-test:
  win32: test the decoder contract and the prediction width path

+++ src/frontend/terminaloverlay.cc
@@ -50,7 +50,7 @@ using namespace Overlay;
    characters would be misjudged as narrow (U+1F600 -> 0xF600 is PUA width 1). */
 static int wcwidth_scalar( char32_t ch )
 {
-  return mosh_win32_wcwidth_scalar( static_cast<uint32_t>( ch ) );
+  return mosh_win32_wcwidth_scalar( ch );
 }
 #else
 static int wcwidth_scalar( char32_t ch )
